+++ index.js
@@ -15,9 +15,9 @@
 const isPromise = require('is-promise')
 const Layer = require('./lib/layer')
 const { METHODS } = require('node:http')
-const mixin = require('utils-merge')
 const parseUrl = require('parseurl')
 const Route = require('./lib/route')
+const debug = require('debug')('router')
 
 /**
  * Module variables.
@@ -149,6 +149,8 @@ Router.prototype.handle = function handle (req, res, callback) {
   if (!callback) {
     throw new TypeError('argument callback is required')
   }
+
+  debug('dispatching %s %s', req.method, req.url)
 
   let idx = 0
   let methods
@@ -316,6 +318,7 @@ Router.prototype.handle = function handle (req, res, callback) {
 
       // Trim off the part of the url that matches the route
       // middleware (.use stuff) needs to have the path stripped
+      debug('trim prefix (%s) from url %s', layerPath, req.url)
       removed = layerPath
       req.url = protohost + req.url.slice(protohost.length + removed.length)
 
@@ -330,6 +333,8 @@ Router.prototype.handle = function handle (req, res, callback) {
         ? removed.substring(0, removed.length - 1)
         : removed)
     }
+
+    debug('%s %s : %s', layer.name, layerPath, req.originalUrl)
 
     if (layerError) {
       layer.handleError(layerError, req, res, next)
@@ -388,6 +393,8 @@ Router.prototype.use = function use (handler) {
     }
 
     // add the middleware
+    debug('use %o %s', path, fn.name || '<anonymous>')
+
     const layer = new Layer(path, {
       sensitive: this.caseSensitive,
       strict: false,
@@ -527,11 +534,11 @@ function mergeParams (params, parent) {
   }
 
   // make copy of parent for base
-  const obj = mixin({}, parent)
+  const obj = Object.assign({}, parent)
 
   // simple non-numeric merging
   if (!(0 in params) || !(0 in parent)) {
-    return mixin(obj, params)
+    return Object.assign(obj, params)
   }
 
   let i = 0
@@ -557,7 +564,7 @@ function mergeParams (params, parent) {
     }
   }
 
-  return mixin(obj, params)
+  return Object.assign(obj, params)
 }
 
 /**