+++ HISTORY.md
@@ -2,9 +2,19 @@ unreleased
 ==================
 
 * Remove `setImmediate` support check
+* Restore `debug` dependency
+
+2.1.0 / 2025-02-10
+==================
+
+* Updated `engines` field to Node@18 or higher
 * Remove `Object.setPrototypeOf` polyfill
 * Use `Array.flat` instead of `array-flatten` package
 * Replace `methods` dependency with standard library
+* deps: parseurl@^1.3.3
+* deps: is-promise@^4.0.0
+* Replace `utils-merge` dependency with `Object.assign`
+* deps: Remove unused dep `after`
 
 2.0.0 / 2024-09-09
 ==================
@@ -13,7 +23,7 @@ unreleased
 * deps: path-to-regexp@^8.0.0
   - Drop support for partial capture group `router.route('/user(s?)/:user/:op')` but still have optional non-capture `/user{s}/:user/:op`
   - `:name?` becomes `{:name}`
-  - `:name*` becomes `*name`. 
+  - `:name*` becomes `*name`.
   - The splat change also changes splat from strings to an array of strings
   - Optional splats become `{*name}`
   - `:name+` becomes `*name` and thus equivalent to `*name` so I dropped those tests